Testing utilities WIP and reorganization
