+++ avs/incredible-squaring-avs/src/operator.rs
@@ -419,820 +419,3 @@ impl<T: Config, I: OperatorInfoServiceTrait> Operator<T, I> {
         Ok(signed_task_response)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use alloy_primitives::{address, Address, Uint};
-    use alloy_provider::Provider;
-    use alloy_provider::ProviderBuilder;
-    use alloy_signer_local::PrivateKeySigner;
-    use alloy_transport_ws::WsConnect;
-    use k256::ecdsa::VerifyingKey;
-    use k256::elliptic_curve::SecretKey;
-    use std::path::Path;
-    use std::time::Duration;
-    use std::vec;
-    //, Bytes, U256};
-
-    static BLS_PASSWORD: &str = "BLS_PASSWORD";
-    static ECDSA_PASSWORD: &str = "ECDSA_PASSWORD";
-    static TASK_RESPONSE_WINDOW_BLOCK: u32 = 10;
-    static TASK_DURATION_BLOCKS: u32 = 0;
-    // static QUORUM_THRESHOLD_PERCENTAGE: U256 = U256::from(100);
-    static AGGREGATOR_ADDR: Address = address!("a0Ee7A142d267C1f36714E4a8F75612F20a79720");
-    static TASK_GENERATOR_ADDR: Address = address!("a0Ee7A142d267C1f36714E4a8F75612F20a79720");
-
-    use crate::avs;
-    use alloy::signers::Signer;
-    use alloy_primitives::Keccak256;
-    use alloy_provider::network::{TransactionBuilder, TxSigner};
-    use alloy_rpc_types_eth::BlockId;
-    use alloy_sol_types::abi::Encoder;
-    use alloy_sol_types::private::SolTypeValue;
-    use anvil::spawn;
-    use ark_bn254::Fq as F;
-    use ark_bn254::{Fr, G1Affine, G2Affine, G2Projective};
-    use avs::IncredibleSquaringServiceManager;
-    use eigen_contracts::RegistryCoordinator::{OperatorSetParam, StrategyParams};
-    use eigen_contracts::*;
-    // use ethabi::ethereum_types::H160;
-    // use ethabi::{encode, Token};
-    use alloy_sol_types::abi::{self, token::*};
-    use alloy_sol_types::Word;
-    use eigen_utils::encode_params;
-    use eigen_utils::test_utils::abi::encode_with_selector;
-    use gadget_common::subxt_signer::bip39::rand_core::OsRng;
-    use url::Url;
-
-    struct ContractAddresses {
-        pub service_manager: Address,
-        pub registry_coordinator: Address,
-        pub operator_state_retriever: Address,
-        pub delegation_manager: Address,
-        pub avs_directory: Address,
-        pub operator: Address,
-    }
-
-    async fn run_anvil_testnet() -> ContractAddresses {
-        // Initialize the logger
-        let _ = env_logger::try_init();
-
-        let (api, mut handle) = spawn(anvil::NodeConfig::test().with_port(33125)).await;
-        api.anvil_auto_impersonate_account(true).await.unwrap();
-        // let http_provider = handle.http_provider();
-        // let ws_provider = handle.ws_provider();
-
-        let _http_provider = ProviderBuilder::new()
-            .on_http(Url::parse(&handle.http_endpoint()).unwrap())
-            .root()
-            .clone();
-        // todo: http_provider is unused
-
-        // let provider = ProviderBuilder::new().on_ws(WsConnect::new(handle.ws_endpoint())).await.unwrap();
-
-        let provider = ProviderBuilder::new()
-            .on_builtin(&handle.ws_endpoint())
-            .await
-            .unwrap();
-
-        let accounts = handle.dev_wallets().collect::<Vec<_>>();
-        let from = accounts[0].address();
-        let _to = accounts[1].address();
-
-        let _amount = handle
-            .genesis_balance()
-            .checked_div(U256::from(2u64))
-            .unwrap();
-
-        let _gas_price = provider.get_gas_price().await.unwrap();
-
-        // Empty address for initial deployment of all contracts
-        let empty_address = Address::default();
-
-        // let strategy_manager_addr = address!("Dc64a140Aa3E981100a9becA4E685f962f0cF6C9");
-        // let delegation_manager_addr = address!("Cf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9");
-        // let avs_directory_addr = address!("5FC8d32690cc91D4c39d9d3abcBD16989F875707");
-        // let proxy_admin_addr = address!("5FbDB2315678afecb367f032d93F642f64180aa3");
-        // let pauser_registry_addr = address!("e7f1725E7734CE288F8367e1Bb143E90bb3F0512");
-        // let base_strategy_addr = address!("322813Fd9A801c5507c9de605d63CEA4f2CE6c44");
-
-        // Deploy Eigenlayer Contracts
-        // let strategy_manager_addr = address!("Dc64a140Aa3E981100a9becA4E685f962f0cF6C9");
-        // let delegation_manager_addr = address!("Cf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9");
-        // let avs_directory_addr = address!("5FC8d32690cc91D4c39d9d3abcBD16989F875707");
-        // let proxy_admin_addr = address!("5FbDB2315678afecb367f032d93F642f64180aa3");
-        // let pauser_registry_addr = address!("e7f1725E7734CE288F8367e1Bb143E90bb3F0512");
-        // let base_strategy_addr = address!("322813Fd9A801c5507c9de605d63CEA4f2CE6c44");
-
-        // let istrategy_manager = IStrategyManager::new(strategy_manager_addr, provider.clone());
-        // let idelegation_manager =
-        //     IDelegationManager::new(delegation_manager_addr, provider.clone());
-        // let iavs_directory = IAVSDirectory::new(avs_directory_addr, provider.clone());
-        // let proxy_admin = ProxyAdmin::new(proxy_admin_addr, provider.clone());
-        // let pauser_registry = PauserRegistry::new(pauser_registry_addr, provider.clone());
-        // let base_strategy = StrategyBaseTVLLimits::new(base_strategy_addr, provider.clone());
-
-        let istrategy_manager = IStrategyManager::deploy(provider.clone()).await.unwrap();
-        let &strategy_manager_addr = istrategy_manager.address();
-
-        let idelegation_manager = IDelegationManager::deploy(provider.clone()).await.unwrap();
-        let &delegation_manager_addr = idelegation_manager.address();
-
-        let iavs_directory = IAVSDirectory::deploy(provider.clone()).await.unwrap();
-        let &avs_directory_addr = iavs_directory.address();
-
-        let proxy_admin = ProxyAdmin::deploy(provider.clone()).await.unwrap();
-        let &proxy_admin_addr = proxy_admin.address();
-
-        let pauser_registry = PauserRegistry::deploy(provider.clone()).await.unwrap();
-        let &pauser_registry_addr = pauser_registry.address();
-
-        let base_strategy = StrategyBaseTVLLimits::deploy(provider.clone(), Default::default())
-            .await
-            .unwrap();
-        let &base_strategy_addr = base_strategy.address();
-
-        let erc20_mock = ERC20Mock::deploy(provider.clone()).await.unwrap();
-        let &erc20_mock_addr = erc20_mock.address();
-
-        let ierc20 = IERC20::new(erc20_mock_addr, provider.clone());
-        let &ierc20_addr = ierc20.address();
-
-        // Function with signature initialize(uint256,uint256,address,address) and selector 0x019e2729.
-        let function_signature = "initialize(uint256,uint256,address,address)";
-        // let params = vec![
-        //     1.tokenize(),
-        //     100.tokenize(),
-        //     ierc20_addr.tokenize(),
-        //     pauser_registry_addr.tokenize(),
-        //     // WordToken(Word::from(1)),
-        //     // WordToken(Word::from(100)),
-        //     // WordToken(Word::from(ierc20_addr.as_slice())),
-        //     // WordToken(Word::from(pauser_registry_addr.as_slice())),
-        // ];
-        // let encoded_data = encode_with_selector(function_signature, params);
-        // let encoded_data = encode_params!(function_signature, 1, 100, ierc20_addr, pauser_registry_addr);
-
-        let mut hasher = Keccak256::new();
-        hasher.update(function_signature);
-        let function_selector = &hasher.finalize()[..4];
-        let hex_selector = hex::encode(function_selector);
-        log::info!("Function selector as hex: {:?}", hex_selector);
-        let mut data = Vec::from(function_selector);
-        // let encoded_param = SolValue::abi_encode(&1);
-        // data.extend_from_slice(&encoded_param);
-        // let encoded_param = SolValue::abi_encode(&100);
-        // data.extend_from_slice(&encoded_param);
-        // let encoded_param = SolValue::abi_encode(&ierc20_addr);
-        // data.extend_from_slice(&encoded_param);
-        // let encoded_param = SolValue::abi_encode(&pauser_registry_addr);
-        // data.extend_from_slice(&encoded_param);
-        let token = 1.tokenize();
-        let encoded_param = abi::encode(&token);
-        data.extend(&encoded_param);
-        let token = 100.tokenize();
-        let encoded_param = abi::encode(&token);
-        data.extend(&encoded_param);
-        let token = ierc20_addr.tokenize();
-        let encoded_param = abi::encode(&token);
-        data.extend(&encoded_param);
-        let token = pauser_registry_addr.tokenize();
-        let encoded_param = abi::encode(&token);
-        data.extend(&encoded_param);
-
-        let encoded_data = alloy_primitives::Bytes::from(data);
-
-        let strategy_proxy = TransparentUpgradeableProxy::deploy(
-            provider.clone(),
-            base_strategy_addr,
-            proxy_admin_addr,
-            encoded_data,
-        )
-        .await
-        .unwrap();
-        let &strategy_proxy_addr = strategy_proxy.address();
-
-        let erc20_mock_strategy =
-            StrategyBaseTVLLimits::deploy(provider.clone(), strategy_proxy_addr)
-                .await
-                .unwrap();
-        let &erc20_mock_strategy_addr = erc20_mock_strategy.address();
-
-        let strategies = vec![erc20_mock_strategy_addr];
-
-        let add_strategies = istrategy_manager
-            .addStrategiesToDepositWhitelist(strategies.clone(), vec![false])
-            .send()
-            .await
-            .unwrap()
-            .get_receipt()
-            .await
-            .unwrap();
-        println!("Add Strategies Receipt: {:?}", add_strategies);
-
-        // Deploy Incredible Squaring Contracts
-        let number_of_strategies = strategies.len();
-
-        let incredible_squaring_proxy_admin = ProxyAdmin::deploy(provider.clone()).await.unwrap();
-        let &incredible_squaring_proxy_admin_addr = incredible_squaring_proxy_admin.address();
-
-        let pausers = vec![
-            address!("f39Fd6e51aad88F6F4ce6aB8827279cffFb92266"),
-            address!("f39Fd6e51aad88F6F4ce6aB8827279cffFb92266"),
-        ];
-
-        let incredible_squaring_pauser_registry_addr =
-            PauserRegistry::deploy_builder(provider.clone())
-                .from(from)
-                .send()
-                .await
-                .unwrap()
-                .get_receipt()
-                .await
-                .unwrap()
-                .contract_address
-                .unwrap();
-        let incredible_squaring_pauser_registry =
-            PauserRegistry::new(incredible_squaring_pauser_registry_addr, provider.clone());
-        let is_pauser = incredible_squaring_pauser_registry
-            .isPauser(address!("f39Fd6e51aad88F6F4ce6aB8827279cffFb92266"))
-            .call()
-            .await
-            .unwrap();
-        println!("Is Dev Account 0 Pauser: {:?}", is_pauser._0);
-
-        let empty_contract = EmptyContract::deploy(provider.clone()).await.unwrap();
-        let &empty_contract_addr = empty_contract.address();
-
-        let incredible_squaring_service_manager = IncredibleSquaringServiceManager::new(
-            TransparentUpgradeableProxy::deploy(
-                provider.clone(),
-                empty_contract_addr,
-                incredible_squaring_proxy_admin_addr,
-                Bytes::from(""),
-            )
-            .await
-            .unwrap()
-            .address()
-            .clone(),
-            provider.clone(),
-        );
-        let &incredible_squaring_service_manager_addr =
-            incredible_squaring_service_manager.address();
-
-        let incredible_squaring_task_manager = IncredibleSquaringTaskManager::new(
-            TransparentUpgradeableProxy::deploy(
-                provider.clone(),
-                empty_contract_addr,
-                incredible_squaring_proxy_admin_addr,
-                Bytes::from(""),
-            )
-            .await
-            .unwrap()
-            .address()
-            .clone(),
-            provider.clone(),
-        );
-        let &incredible_squaring_task_manager_addr = incredible_squaring_task_manager.address();
-
-        let registry_coordinator = RegistryCoordinator::new(
-            TransparentUpgradeableProxy::deploy(
-                provider.clone(),
-                empty_contract_addr,
-                incredible_squaring_proxy_admin_addr,
-                Bytes::from(""),
-            )
-            .await
-            .unwrap()
-            .address()
-            .clone(),
-            provider.clone(),
-        );
-        let &registry_coordinator_addr = registry_coordinator.address();
-
-        let bls_apk_registry = BlsApkRegistry::new(
-            TransparentUpgradeableProxy::deploy(
-                provider.clone(),
-                empty_contract_addr,
-                incredible_squaring_proxy_admin_addr,
-                Bytes::from(""),
-            )
-            .await
-            .unwrap()
-            .address()
-            .clone(),
-            provider.clone(),
-        );
-        let &bls_apk_registry_addr = bls_apk_registry.address();
-
-        let bls_apk_registry = IBlsApkRegistry::new(
-            TransparentUpgradeableProxy::deploy(
-                provider.clone(),
-                empty_contract_addr,
-                incredible_squaring_proxy_admin_addr,
-                Bytes::from(""),
-            )
-            .await
-            .unwrap()
-            .address()
-            .clone(),
-            provider.clone(),
-        );
-        let &bls_apk_registry_addr = bls_apk_registry.address();
-
-        let index_registry = IIndexRegistry::new(
-            TransparentUpgradeableProxy::deploy(
-                provider.clone(),
-                empty_contract_addr,
-                incredible_squaring_proxy_admin_addr,
-                Bytes::from(""),
-            )
-            .await
-            .unwrap()
-            .address()
-            .clone(),
-            provider.clone(),
-        );
-        let &index_registry_addr = index_registry.address();
-
-        let stake_registry = IStakeRegistry::new(
-            TransparentUpgradeableProxy::deploy(
-                provider.clone(),
-                empty_contract_addr,
-                incredible_squaring_proxy_admin_addr,
-                Bytes::from(""),
-            )
-            .await
-            .unwrap()
-            .address()
-            .clone(),
-            provider.clone(),
-        );
-        let &stake_registry_addr = stake_registry.address();
-
-        let operator_state_retriever = OperatorStateRetriever::deploy(provider.clone())
-            .await
-            .unwrap();
-        let &operator_state_retriever_addr = operator_state_retriever.address();
-
-        //Now, deploy the implementation contracts using the proxy contracts as inputs
-        let stake_registry_implementation = StakeRegistry::deploy(
-            provider.clone(),
-            registry_coordinator_addr,
-            delegation_manager_addr,
-        )
-        .await
-        .unwrap();
-        let &stake_registry_implementation_addr = stake_registry_implementation.address();
-        let stake_registry_upgrade = incredible_squaring_proxy_admin
-            .upgrade(stake_registry_addr, stake_registry_implementation_addr)
-            .send()
-            .await
-            .unwrap()
-            .get_receipt()
-            .await
-            .unwrap();
-        log::info!(
-            "Stake Registry Upgrade Receipt: {:?}",
-            stake_registry_upgrade
-        );
-
-        let bls_apk_registry_implementation =
-            BlsApkRegistry::deploy(provider.clone(), registry_coordinator_addr)
-                .await
-                .unwrap();
-        let &bls_apk_registry_implementation_addr = bls_apk_registry_implementation.address();
-        let bls_apk_registry_upgrade = incredible_squaring_proxy_admin
-            .upgrade(bls_apk_registry_addr, bls_apk_registry_implementation_addr)
-            .send()
-            .await
-            .unwrap()
-            .get_receipt()
-            .await
-            .unwrap();
-        log::info!(
-            "Bls Apk Registry Upgrade Receipt: {:?}",
-            bls_apk_registry_upgrade
-        );
-
-        let index_registry_implementation =
-            IndexRegistry::deploy(provider.clone(), registry_coordinator_addr)
-                .await
-                .unwrap();
-        let &index_registry_implementation_addr = index_registry_implementation.address();
-        let index_registry_upgrade = incredible_squaring_proxy_admin
-            .upgrade(index_registry_addr, index_registry_implementation_addr)
-            .send()
-            .await
-            .unwrap()
-            .get_receipt()
-            .await
-            .unwrap();
-        log::info!(
-            "Index Registry Upgrade Receipt: {:?}",
-            index_registry_upgrade
-        );
-
-        let registry_coordinator_implementation = RegistryCoordinator::deploy(
-            provider.clone(),
-            incredible_squaring_service_manager_addr,
-            stake_registry_addr,
-            bls_apk_registry_addr,
-            index_registry_addr,
-        )
-        .await
-        .unwrap();
-        let &registry_coordinator_implementation_addr =
-            registry_coordinator_implementation.address();
-
-        let number_of_quorums = 1;
-        // For each quorum we want to set up, we must define QuorumOperatorSetParam, minimumStakeForQuorum, and strategyParams
-        let mut quorum_operator_set_params = Vec::<OperatorSetParam>::new();
-        for i in 0..number_of_quorums {
-            log::info!("Deploying quorum {}", i);
-            quorum_operator_set_params.push(OperatorSetParam {
-                maxOperatorCount: 10000,
-                kickBIPsOfOperatorStake: 15000,
-                kickBIPsOfTotalStake: 100,
-            });
-        }
-        // Set to 0 for each quorum
-        let quorums_minimum_stake = Vec::<u128>::new();
-        let mut quorums_strategy_params = Vec::<Vec<StrategyParams>>::new();
-
-        for j in 0..number_of_quorums {
-            quorums_strategy_params.push(Vec::<StrategyParams>::new());
-            for k in 0..number_of_strategies {
-                quorums_strategy_params[j][k] = StrategyParams {
-                    strategy: strategies[j],
-                    multiplier: 1,
-                }
-            }
-        }
-
-        // Function with signature initialize(address,address,address,address,uint256,(uint32,uint16,uint16)[],uint96[],(address,uint96)[][]) and selector 0xdd8283f3.
-        let function_signature = "initialize(address,address,address,address,uint256,(uint32,uint16,uint16)[],uint96[],(address,uint96)[][])";
-
-        let encoded_data = encode_params!(
-            function_signature,
-            pausers[0],
-            pausers[0],
-            pausers[0],
-            pausers[1],
-            0,
-            quorum_operator_set_params,
-            quorums_minimum_stake,
-            quorums_strategy_params
-        );
-        let registry_coordinator_upgrade = incredible_squaring_proxy_admin
-            .upgradeAndCall(
-                registry_coordinator_addr,
-                registry_coordinator_implementation_addr,
-                alloy_primitives::Bytes::from(encoded_data),
-            )
-            .send()
-            .await
-            .unwrap()
-            .get_receipt()
-            .await
-            .unwrap();
-        log::info!(
-            "Registry Coordinator Upgrade Receipt: {:?}",
-            registry_coordinator_upgrade
-        );
-
-        let incredible_squaring_service_manager_implementation =
-            IncredibleSquaringServiceManager::deploy(
-                provider.clone(),
-                avs_directory_addr,
-                registry_coordinator_addr,
-                stake_registry_addr,
-                incredible_squaring_task_manager_addr,
-            )
-            .await
-            .unwrap();
-        let &incredible_squaring_service_manager_implementation_addr =
-            incredible_squaring_service_manager_implementation.address();
-        let incredible_squaring_service_manager_upgrade = incredible_squaring_proxy_admin
-            .upgrade(
-                incredible_squaring_service_manager_addr,
-                incredible_squaring_service_manager_implementation_addr,
-            )
-            .send()
-            .await
-            .unwrap()
-            .get_receipt()
-            .await
-            .unwrap();
-        log::info!(
-            "Incredible Squaring Service Manager Upgrade Receipt: {:?}",
-            incredible_squaring_service_manager_upgrade
-        );
-
-        // Function with signature initialize(address,address,address,address) and selector 0xf8c8765e.
-        let function_signature = "initialize(address,address,address,address)";
-        // let params = vec![
-        //     pauser_registry_addr.tokenize(),
-        //     pausers[0].tokenize(),
-        //     AGGREGATOR_ADDR.tokenize(),
-        //     TASK_GENERATOR_ADDR.tokenize(),
-        // ];
-        // let encoded_data =
-        //     eigen_utils::test_utils::abi::encode_with_selector(function_signature, params);
-
-        let encoded_data = encode_params!(
-            function_signature,
-            pauser_registry_addr,
-            pausers[0],
-            AGGREGATOR_ADDR,
-            TASK_GENERATOR_ADDR
-        );
-        let incredible_squaring_task_manager_implementation =
-            IncredibleSquaringTaskManager::deploy(
-                provider.clone(),
-                registry_coordinator_addr,
-                TASK_RESPONSE_WINDOW_BLOCK,
-            )
-            .await
-            .unwrap();
-        let &incredible_squaring_task_manager_implementation_addr =
-            incredible_squaring_task_manager_implementation.address();
-        let incredible_squaring_task_manager_upgrade = incredible_squaring_proxy_admin
-            .upgradeAndCall(
-                incredible_squaring_task_manager_addr,
-                incredible_squaring_service_manager_implementation_addr,
-                alloy_primitives::Bytes::from(encoded_data),
-            )
-            .send()
-            .await
-            .unwrap()
-            .get_receipt()
-            .await
-            .unwrap();
-        log::info!(
-            "Incredible Squaring Task Manager Upgrade Receipt: {:?}",
-            incredible_squaring_task_manager_upgrade
-        );
-
-        log::info!("ERC20MOCK ADDRESS: {:?}", erc20_mock_addr);
-        log::info!("ERC20MOCK STRATEGY ADDRESS: {:?}", erc20_mock_strategy_addr);
-        log::info!(
-            "INCREDIBLE SQUARING TASK MANAGER ADDRESS: {:?}",
-            incredible_squaring_task_manager_addr
-        );
-        log::info!(
-            "INCREDIBLE SQUARING TASK MANAGER IMPLEMENTATION ADDRESS: {:?}",
-            incredible_squaring_task_manager_implementation_addr
-        );
-        log::info!(
-            "INCREDIBLE SQUARING SERVICE MANAGER ADDRESS: {:?}",
-            incredible_squaring_service_manager_addr
-        );
-        log::info!(
-            "INCREDIBLE SQUARING SERVICE MANAGER IMPLEMENTATION ADDRESS: {:?}",
-            incredible_squaring_service_manager_implementation_addr
-        );
-        log::info!(
-            "REGISTRY COORDINATOR ADDRESS: {:?}",
-            registry_coordinator_addr
-        );
-        log::info!(
-            "REGISTRY COORDINATOR IMPLEMENTATION ADDRESS: {:?}",
-            registry_coordinator_implementation_addr
-        );
-        log::info!(
-            "OPERATOR STATE RETRIEVER ADDRESS: {:?}",
-            operator_state_retriever_addr
-        );
-
-        // let _block = provider
-        //     .get_block(BlockId::latest(), false.into())
-        //     .await
-        //     .unwrap()
-        //     .unwrap();
-        //
-        // api.anvil_set_auto_mine(true).await.unwrap();
-        // let run_testnet = async move {
-        //     let serv = handle.servers.pop().unwrap();
-        //     let res = serv.await.unwrap();
-        //     res.unwrap();
-        // };
-        // let spawner_task_manager_address = task_manager_addr.clone();
-        // // let spawner_provider = provider.clone();
-        // let spawner_provider = provider;
-        // let task_spawner = async move {
-        //     let manager = IncredibleSquaringTaskManager::new(
-        //         spawner_task_manager_address,
-        //         spawner_provider.clone(),
-        //     );
-        //     loop {
-        //         api.mine_one().await;
-        //         log::info!("About to create new task");
-        //         tokio::time::sleep(std::time::Duration::from_millis(5000)).await;
-        //         let result = manager
-        //             .createNewTask(U256::from(2), 100u32, Bytes::from("0"))
-        //             .send()
-        //             .await
-        //             .unwrap()
-        //             .watch()
-        //             .await
-        //             .unwrap();
-        //         api.mine_one().await;
-        //         log::info!("Created new task: {:?}", result);
-        //         // let latest_task = manager.latestTaskNum().call().await.unwrap()._0;
-        //         // log::info!("Latest task: {:?}", latest_task);
-        //         // let task_hash = manager.allTaskHashes(latest_task).call().await.unwrap()._0;
-        //         // log::info!("Task info: {:?}", task_hash);
-        //     }
-        // };
-        // tokio::spawn(run_testnet);
-        // tokio::spawn(task_spawner);
-
-        ContractAddresses {
-            service_manager: incredible_squaring_service_manager_implementation_addr,
-            registry_coordinator: registry_coordinator_implementation_addr,
-            operator_state_retriever: operator_state_retriever_addr,
-            delegation_manager: delegation_manager_addr,
-            avs_directory: avs_directory_addr,
-            operator: from,
-        }
-    }
-
-    #[tokio::test]
-    async fn test_anvil() {
-        env_logger::init();
-
-        // Runs new Anvil Testnet - used for deploying programmatically in rust
-        let contract_addresses = run_anvil_testnet().await;
-
-        // // Runs saved Anvil Testnet - loads from saved chain state JSON file
-        // let chain = eigen_utils::test_utils::local_chain::LocalEvmChain::new_with_chain_state(
-        //     31337,
-        //     String::from("eigen-testnet"),
-        //     Path::new("../../eigen-utils/saved-anvil-state.json"),
-        //     Some(8545u16),
-        // );
-        // let chain_id = chain.chain_id();
-        // let chain_name = chain.name();
-        // println!("chain_id: {:?}", chain_id);
-        // println!("chain_name: {:?}", chain_name);
-
-        // let account_one = address!("f39Fd6e51aad88F6F4ce6aB8827279cffFb92266");
-        // let account_two = address!("70997970C51812dc3A010C7d01b50e0d17dc79C8");
-
-        // let contract_addresses = ContractAddresses {
-        //     service_manager: address!("84eA74d481Ee0A5332c457a4d796187F6Ba67fEB"),
-        //     registry_coordinator: address!("a82fF9aFd8f496c3d6ac40E2a0F282E47488CFc9"),
-        //     operator_state_retriever: address!("95401dc811bb5740090279Ba06cfA8fcF6113778"),
-        //     delegation_manager: address!("Cf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9"),
-        //     avs_directory: address!("5FC8d32690cc91D4c39d9d3abcBD16989F875707"),
-        //     operator: account_two,
-        // };
-
-        // Implementation version of addresses
-        // let contract_addresses = ContractAddresses {
-        //     service_manager: address!("84eA74d481Ee0A5332c457a4d796187F6Ba67fEB"),
-        //     registry_coordinator: address!("9d4454B023096f34B160D6B654540c56A1F81688"),
-        //     operator_state_retriever: address!("95401dc811bb5740090279Ba06cfA8fcF6113778"),
-        //     delegation_manager: address!("B7f8BC63BbcaD18155201308C8f3540b07f84F5e"),
-        //     avs_directory: address!("0DCd1Bf9A1b36cE34237eEaFef220932846BCD82"),
-        //     operator: address!("f39Fd6e51aad88F6F4ce6aB8827279cffFb92266"),
-        // };
-
-        let http_endpoint = "http://127.0.0.1:8545";
-        let ws_endpoint = "ws://127.0.0.1:8545";
-        let node_config = NodeConfig {
-            node_api_ip_port_address: "127.0.0.1:9808".to_string(),
-            eth_rpc_url: http_endpoint.to_string(),
-            eth_ws_url: ws_endpoint.to_string(),
-            bls_private_key_store_path: "./keystore/bls".to_string(),
-            ecdsa_private_key_store_path: "./keystore/ecdsa".to_string(),
-            incredible_squaring_service_manager_addr: contract_addresses
-                .service_manager
-                .to_string(),
-            avs_registry_coordinator_addr: contract_addresses.registry_coordinator.to_string(),
-            operator_state_retriever_addr: contract_addresses.operator_state_retriever.to_string(),
-            eigen_metrics_ip_port_address: "127.0.0.1:9100".to_string(),
-            delegation_manager_addr: contract_addresses.delegation_manager.to_string(),
-            avs_directory_addr: contract_addresses.avs_directory.to_string(),
-            operator_address: contract_addresses.operator.to_string(),
-            enable_metrics: false,
-            enable_node_api: false,
-            server_ip_port_address: "".to_string(),
-        };
-
-        let operator_info_service = OperatorInfoService {};
-
-        let hex_key =
-            hex::decode("59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d")
-                .unwrap();
-        let secret_key = SecretKey::from_slice(&hex_key).unwrap();
-        let signing_key = SigningKey::from(secret_key.clone());
-        let signer = EigenGadgetSigner {
-            signer: PrivateKeySigner::from_signing_key(signing_key),
-        };
-
-        println!("Creating HTTP Provider...");
-
-        let http_provider = ProviderBuilder::new()
-            .with_recommended_fillers()
-            .on_http(http_endpoint.parse().unwrap())
-            .root()
-            .clone()
-            .boxed();
-
-        println!("Creating WS Provider...");
-
-        let ws_provider = ProviderBuilder::new()
-            .with_recommended_fillers()
-            .on_ws(WsConnect::new(ws_endpoint))
-            .await
-            .unwrap()
-            .root()
-            .clone()
-            .boxed();
-
-        println!("Now setting up Operator!");
-
-        let operator = Operator::<NodeConfig, OperatorInfoService>::new_from_config(
-            node_config.clone(),
-            EigenGadgetProvider {
-                provider: http_provider,
-            },
-            EigenGadgetProvider {
-                provider: ws_provider,
-            },
-            operator_info_service,
-            signer,
-        )
-        .await
-        .unwrap();
-
-        operator.start().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_start_chain_from_state() {
-        env_logger::init();
-
-        let chain = eigen_utils::test_utils::local_chain::LocalEvmChain::new_with_chain_state(
-            31337,
-            String::from("eigen-testnet"),
-            Path::new("../../eigen-utils/eigen-gadget-anvil-state.json"),
-            Some(8545u16),
-        );
-        let chain_id = chain.chain_id();
-        let chain_name = chain.name();
-        println!("chain_id: {:?}", chain_id);
-        println!("chain_name: {:?}", chain_name);
-        let addresses = chain.addresses();
-        println!("addresses: {:?}", addresses);
-        tokio::time::sleep(Duration::from_secs(5)).await;
-        println!("Now shutting down...");
-        chain.shutdown();
-    }
-
-    #[tokio::test]
-    async fn test_generate_keys() {
-        env_logger::init();
-
-        // ---------------- BLS ----------------
-        let bls_pair = KeyPair::gen_random().unwrap();
-        bls_pair
-            .save_to_file("./keystore/bls", BLS_PASSWORD)
-            .unwrap();
-        let bls_keys = KeyPair::read_private_key_from_file("./keystore/bls", BLS_PASSWORD).unwrap();
-        assert_eq!(bls_pair.priv_key, bls_keys.priv_key);
-        assert_eq!(bls_pair.pub_key, bls_keys.pub_key);
-
-        //---------------- ECDSA ----------------
-        // First Account
-        // let hex_key =
-        //     hex::decode("ac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80")
-        //         .unwrap();
-
-        // Second Account
-        let hex_key =
-            hex::decode("59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d")
-                .unwrap();
-
-        let secret_key = SecretKey::from_slice(&hex_key).unwrap();
-        let signing_key = SigningKey::from(secret_key.clone());
-        let public_key = secret_key.public_key();
-        let verifying_key = VerifyingKey::from(public_key);
-        eigen_utils::crypto::ecdsa::write_key("./keystore/ecdsa", &secret_key, ECDSA_PASSWORD)
-            .unwrap();
-
-        let read_ecdsa_secret_key =
-            eigen_utils::crypto::ecdsa::read_key("./keystore/ecdsa", ECDSA_PASSWORD).unwrap();
-        let read_ecdsa_public_key = read_ecdsa_secret_key.public_key();
-        let read_ecdsa_signing_key = SigningKey::from(&read_ecdsa_secret_key);
-        let read_ecdsa_verifying_key = VerifyingKey::from(&read_ecdsa_signing_key);
-
-        assert_eq!(secret_key, read_ecdsa_secret_key);
-        assert_eq!(public_key, read_ecdsa_public_key);
-        assert_eq!(signing_key, read_ecdsa_signing_key);
-        assert_eq!(verifying_key, read_ecdsa_verifying_key);
-    }
-}
