+++ eigensdk-rs/src/lib.rs
@@ -1,5 +1,5 @@
 pub use eigen_contracts;
-pub use tangle_avs;
-pub use incredible_squaring_avs;
 pub use eigen_utils;
 pub use fireblocks_client;
+pub use incredible_squaring_avs;
+pub use tangle_avs;
