+++ test-utils/src/anvil/mod.rs
@@ -1,4 +1,5 @@
+pub mod abi;
 pub mod anvil;
 pub mod local_chain;
 mod random_port;
-pub mod abi;
+pub mod testnet;
