+++ test-utils/src/anvil/testnet/mod.rs
@@ -1,0 +1,1 @@
+pub mod incredible_squaring;
